fix bookmarks state
