+++ src/components/detail-book/BookOverview.module.scss
@@ -1,0 +1,141 @@
+.book {
+  &__container {
+    display: flex;
+    width: 70%;
+    margin: auto;
+  }
+
+  &__cover {
+    // max-width: 38%;
+    min-width: 12rem;
+
+    height: 100%;
+    flex-basis: 20%;
+    margin-right: 100px;
+    position: relative;
+
+    img {
+      border: 10px solid var(--color-2);
+      border-radius: 10px;
+      width: 98%;
+      height: 100%;
+      object-fit: contain;
+      display: block;
+      transition: all 0.2s;
+    }
+
+    &:hover {
+      img {
+        transform: scale(1.01);
+      }
+    }
+  }
+
+  &__bookmark {
+    position: absolute;
+    outline: none;
+    border: none;
+    top: -20px;
+    left: 10%;
+    z-index: 500;
+    background-color: transparent;
+    transition: all 0.2s;
+
+    svg {
+      width: 40px;
+      fill: var(--color-2);
+      color: var(--color-3);
+    }
+
+    &:hover {
+      transform: scale(1.05);
+      svg {
+        fill: var(--color-text-2);
+      }
+    }
+
+    &--active {
+      svg {
+        fill: var(--color-text-2);
+      }
+    }
+  }
+
+  &__content {
+    flex-basis: 70%;
+  }
+
+  &__title {
+    font-family: var(--font-heading);
+    font-weight: 600;
+    color: var(--color-text-1);
+    font-size: 32px;
+    line-height: 1;
+  }
+
+  &__authors {
+    margin-top: -5px;
+    font-weight: 500;
+    color: var(--color-text-2);
+  }
+
+  &__summary--title {
+    font-size: 20px;
+    font-weight: 500;
+    text-transform: capitalize;
+    margin-top: 30px;
+  }
+  &__summary--text {
+    color: var(--colot-text-2);
+    font-size: 16px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: 1.5;
+    margin: 0;
+    padding: 0;
+  }
+
+  &__additional-info {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    margin-top: 30px;
+
+    li {
+      display: grid;
+      grid-auto-flow: column;
+      grid-template-columns: 30% 1fr;
+      align-items: start;
+    }
+    .title {
+      font-size: 18px;
+      font-weight: 500;
+      text-transform: capitalize;
+    }
+    .text {
+      justify-self: start;
+    }
+
+    a {
+      font-family: inherit;
+      color: inherit;
+      text-decoration: none;
+
+      &:hover {
+        color: var(--color-text-1);
+        text-decoration: underline;
+        cursor: pointer;
+      }
+    }
+  }
+
+  .error {
+    font-size: 20px;
+    text-transform: capitalize;
+    color: var(--color-error);
+    text-align: center;
+  }
+}
