+++ src/components/detail-book/RecommendedBooks.tsx
@@ -3,16 +3,19 @@
 import React from "react";
 import GenreSection from "../home/GenreSection";
 import classes from "./RecommendedBooks.module.scss";
+import { BookDetailsProp } from "../../firebase/services";
 
 interface RecommendedBooksProp {
   authors: string[];
   categories: string[];
   bookmarks: string[];
+  toggleBookmark: (bookDetails: BookDetailsProp, isBookmarked: boolean) => void;
 }
 const RecommendedBooks: React.FC<RecommendedBooksProp> = function ({
   authors,
   categories,
-  bookmarks
+  bookmarks,
+  toggleBookmark
 }) {
   const author = authors?.[0] || "";
   const reqAuthor = author.split(" ").join("+");
@@ -25,6 +28,7 @@ const RecommendedBooks: React.FC<RecommendedBooksProp> = function ({
           name="author"
           title={reqAuthor}
           bookmarks={bookmarks}
+          toggleBookmark={toggleBookmark}
         />
       )}
       {!authors && categories && (
@@ -33,6 +37,7 @@ const RecommendedBooks: React.FC<RecommendedBooksProp> = function ({
           name="genre"
           title={categories?.[0]}
           bookmarks={bookmarks}
+          toggleBookmark={toggleBookmark}
         />
       )}
     </div>
