+++ src/components/home/BookItem.module.scss
@@ -7,7 +7,7 @@
   font-family: inherit;
   padding: 10px 0;
   transition: all 0.2s;
-  width: 100%;
+  // width: 100%;
   // border: 1px solid var(--color-2);
   border-radius: 5px;
   background-color: var(--color-background);
