+++ src/components/home/HeroSection.tsx
@@ -2,15 +2,24 @@
 import InputGroup from "react-bootstrap/InputGroup";
 import Form from "react-bootstrap/Form";
 import CustomButton from "../ui/Button";
-import React, { useState } from "react";
+import React, { useContext, useState } from "react";
 
 import CustomModal from "../ui/CustomModal";
 import classes from "./HeroSection.module.scss";
+import { modifyBookmarks } from "../../firebase/services";
+import { AuthContext } from "../../context/AuthContext";
+import useUser from "../custom-hook/useUser";
+import { BookDetailsProp } from "../../firebase/services";
+import { getBookmarks } from "../../firebase/services";
 
 function HeroSection() {
   const [inputQuery, setInputQuery] = useState("");
 
   const [modalShow, setModalShow] = useState(false);
+
+  const auth = useContext(AuthContext);
+  const user = useUser(auth?.uid);
+  const [bookmarks, setBookmarks] = useState<string[] | []>([]);
 
   //input handler
   const handleInput = async function (e: React.ChangeEvent<HTMLInputElement>) {
@@ -21,6 +30,18 @@ function HeroSection() {
   const handleSearch = async function () {
     if (inputQuery.trim().length <= 0) return;
     setModalShow(true);
+  };
+
+  const toggleBookmarkHandler = async function (
+    bookDetails: BookDetailsProp,
+    isBookmarked: boolean
+  ) {
+    if (!user) return;
+
+    await modifyBookmarks(user?.docId, bookDetails, isBookmarked);
+
+    const req = await getBookmarks(user?.docId);
+    setBookmarks(req);
   };
 
   // console.log(inputQuery, "input");
@@ -49,6 +70,8 @@ function HeroSection() {
           show={modalShow}
           onHide={() => setModalShow(false)}
           query={inputQuery}
+          toggleBookmark={toggleBookmarkHandler}
+          bookmarks={bookmarks}
         />
       )}
     </header>
