+++ src/components/home/ScienceSection.tsx
@@ -1,19 +1,27 @@
 import GenreSection from "./GenreSection";
+import { BookDetailsProp } from "../../firebase/services";
 
-interface BookmarksProp {
+interface ScienceSectionProp {
   bookmarks: string[];
+  toggleBookmark: (
+    bookDetails: BookDetailsProp,
+    isBookmarked: boolean
+  ) => Promise<void>;
 }
 
-const ScienceSection = function (props: BookmarksProp) {
+const ScienceSection = function (props: ScienceSectionProp) {
   const genre = "science";
   // const maxResults = 8;
   // const url = `volumes?q=subject:${genre}&maxResults=${maxResults}&`;
+
+  // console.log(props, "science section");
 
   return (
     <GenreSection
       name="genre"
       title={genre}
       bookmarks={props.bookmarks}
+      toggleBookmark={props.toggleBookmark}
     ></GenreSection>
   );
 };
