+++ src/components/ui/CustomModal.tsx
@@ -6,11 +6,17 @@ import BookItem from "../home/BookItem";
 import { Book } from "../home/GenreSection";
 import classes from "./CustomModal.module.scss";
 import { Spinner } from "react-bootstrap";
+import { BookDetailsProp } from "../../firebase/services";
 
 interface CustomModalProps {
   show: boolean;
   onHide: () => void;
   query: string;
+  bookmarks: string[];
+  toggleBookmark: (
+    bookDetails: BookDetailsProp,
+    isBookmarked: boolean
+  ) => Promise<void>;
 }
 
 const CustomModal: React.FC<CustomModalProps> = function (props) {
@@ -42,6 +48,10 @@ const CustomModal: React.FC<CustomModalProps> = function (props) {
                   title={book?.volumeInfo?.title}
                   author={book?.volumeInfo?.authors}
                   description={book?.volumeInfo?.description}
+                  toggleBookmark={props.toggleBookmark}
+                  bookmarked={props.bookmarks?.some(
+                    (item) => item === book?.id
+                  )}
                 />
               );
             })}
