+++ src/index.scss
@@ -96,7 +96,7 @@ body {
   }
 }
 .modal {
-  top: 10%;
+  top: 30px;
 
   &-body {
     padding: 0;
