+++ src/pages/Homepage.tsx
@@ -2,14 +2,16 @@
 import classes from "./HomePage.module.scss";
 import HeroSection from "../components/home/HeroSection";
 import ScienceSection from "../components/home/ScienceSection";
-import BiographySection from "../components/home/BiographySection";
-import FictionSection from "../components/home/FictionSection";
-import MysterySection from "../components/home/MysterySection";
+// import BiographySection from "../components/home/BiographySection";
+// import FictionSection from "../components/home/FictionSection";
+// import MysterySection from "../components/home/MysterySection";
 import { useState, useEffect, useContext } from "react";
 
 import useUser from "../components/custom-hook/useUser";
 import { AuthContext } from "../context/AuthContext";
 import { getBookmarks } from "../firebase/services";
+import { BookDetailsProp } from "../firebase/services";
+import { modifyBookmarks } from "../firebase/services";
 
 function HomePage() {
   const [bookmarks, setBookmarks] = useState<string[] | []>([]);
@@ -28,11 +30,26 @@ function HomePage() {
     }
   }, [user]);
 
+  const toggleBookmarkHandler = async function (
+    bookDetails: BookDetailsProp,
+    isBookmarked: boolean
+  ) {
+    if (!user) return;
+
+    await modifyBookmarks(user?.docId, bookDetails, isBookmarked);
+
+    const req = await getBookmarks(user?.docId);
+    setBookmarks(req);
+  };
+
   return (
     <section className={classes.homePage}>
       <HeroSection />
       <main>
-        <ScienceSection bookmarks={bookmarks} />
+        <ScienceSection
+          bookmarks={bookmarks}
+          toggleBookmark={toggleBookmarkHandler}
+        />
         {/* <BiographySection bookmarks={bookmarks} />
         <FictionSection bookmarks={bookmarks} />
         <MysterySection bookmarks={bookmarks} /> */}
